refactor: extract product details components into custom hooks and smaller components
