+++ src/components/templates/(marketing)/Index/Product/Details.tsx
@@ -1,202 +1,56 @@
 'use client';
 
-import { useRef, useState } from 'react';
-import { Button } from '@/components/ui/button';
-import { Star } from 'lucide-react';
 import { TProduct } from '@/types';
-import { useAuthStore } from '@/stores/authStore';
 import Colors from './(Details)/Colors';
 import Sizes from './(Details)/Sizes';
-import { toast } from 'sonner';
+import useProductDetails from '@/hooks/(product_details)/useProductDetails';
+import ProductInfo from './(Details)/ProductInfo';
+import QuantitySelector from './(Details)/QuantitySelector';
+import ProductActions from './(Details)/ProductActions';
+import ProductDescription from './(Details)/ProductDescription';
 
 export default function Details({ product }: { product: TProduct }) {
-  const userId = useAuthStore((state) => state.user?.id);
+  const {
+    selectVariant,
+    quantity,
+    onChangeColor,
+    onChangeSize,
+    increaseQuantity,
+    decreaseQuantity,
+  } = useProductDetails(product);
 
-  const [showDescription, setShowDescription] = useState(false);
-  const [selectVariant, setSelectVariant] = useState(
-    product.colors?.[0]?.variants?.[0]
-  );
-  const [quantity, setQuantity] = useState(1);
-
-  const selectedColor = useRef(product.colors?.[0]?.id);
-  const selectedSize = useRef(product.colors?.[0]?.variants?.[0]?.size?.id);
-
-  // TODO: Implement quantity change logic
-  const onChangeColor = (color: number | undefined) => {
-    // *Lưu lại id màu sắc mà người dùng vừa chọn vào biến selectedColor.current.
-    selectedColor.current = color;
-    const colorObj = product.colors?.find((c) => c.id === color);
-    // *Tìm variant phù hợp với size hiện tại
-    let variant = colorObj?.variants?.find(
-      (v) => v.size?.id === selectedSize.current
-    );
-    // *Nếu không tìm thấy variant phù hợp với size hiện tại, lấy variant đầu tiên
-    if (!variant && colorObj?.variants?.length) {
-      variant = colorObj.variants[0];
-      selectedSize.current = variant.size?.id;
-    }
-    setSelectVariant(variant);
+  const handleAddToCart = () => {
+    console.log('Thêm vào giỏ hàng:', {
+      productId: product.id,
+      variantId: selectVariant,
+      quantity,
+    });
+    // Logic thêm vào giỏ hàng
   };
 
-  // TODO: Implement size change logic
-  const onChangeSize = (size: number | undefined) => {
-    // *Lưu lại id size mà người dùng vừa chọn vào biến selectedColor.current.
-    selectedSize.current = size;
-    // *Tìm variant phù hợp với size hiện tại
-    setSelectVariant(
-      product.colors
-        ?.find((c) => c.id === selectedColor.current)
-        ?.variants?.find((v) => v.size?.id === size)
-    );
+  const handleBuyNow = () => {
+    // Logic mua ngay
   };
-
-  const increaseQuantity = () => {
-    if (quantity >= (selectVariant?.quantity ?? 0)) {
-      toast.error('Số lượng vượt quá số hàng có sẵn');
-      return;
-    }
-    setQuantity((prev) => prev + 1);
-  };
-
-  const decreaseQuantity = () => {
-    if (quantity <= 1) {
-      return;
-    }
-    setQuantity((prev) => prev - 1);
-  };
-
-  // console.log(selectedColor.current, selectVariant);
 
   return (
     <div className='space-y-6'>
-      {/* Product Info */}
-      <div className='space-y-1'>
-        <h1 className='text-2xl md:text-3xl lg:text-4xl font-bold text-primary '>
-          {product.name}
-        </h1>
-        <div className='flex gap-2'>
-          <Star className=' size-5 fill-amber-300 text-amber-200' />{' '}
-          <span> {product.averageRating}.0</span>
-          <span className='text-gray-400 ml-4'>
-            Đã bán: {product.soldQuantity || 0}
-          </span>
-        </div>
-        <div className='text-xl md:text-2xl lg:text-3xl font-bold text-primary'>
-          {selectVariant?.price || 0} ₫
-        </div>
-      </div>
+      <ProductInfo product={product} selectVariant={selectVariant} />
 
-      {/* Color Selection */}
       <Colors tColors={product.colors || []} onChangeColor={onChangeColor} />
-
-      {/* Size Selection */}
       <Sizes
         tColor={(product.colors ?? [])[0] ?? []}
         onChangeSize={onChangeSize}
       />
 
-      {/* Quantity */}
-      <div>
-        <label className='block text-sm font-medium text-primary mb-2'>
-          Số lượng:
-        </label>
-        <div className='flex items-center gap-2'>
-          <button
-            onClick={decreaseQuantity}
-            className='w-8 h-8 border border-border rounded flex items-center justify-center hover:bg-[#e8e8e8]'
-          >
-            -
-          </button>
-          <span className='w-12 text-center'>{quantity}</span>
-          <button
-            onClick={increaseQuantity}
-            className='w-8 h-8 border border-border rounded flex items-center justify-center hover:bg-[#e8e8e8]'
-          >
-            +
-          </button>
-        </div>
-      </div>
+      <QuantitySelector
+        quantity={quantity}
+        onDecrease={decreaseQuantity}
+        onIncrease={increaseQuantity}
+      />
 
-      {/* Action Buttons */}
-      <div className='flex gap-2'>
-        <Button className='px-10 py-3 rounded-md font-semibold text-white bg-neutral-800 hover:bg-neutral-700'>
-          Thêm vào giỏ hàng
-        </Button>
-        <Button
-          onClick={() => {
-            console.log(selectVariant, quantity, userId);
-          }}
-          className='px-10 py-3 rounded-md font-semibold text-white bg-blue-800 hover:bg-blue-700'
-        >
-          Mua Ngay
-        </Button>
-      </div>
+      <ProductActions onAddToCart={handleAddToCart} onBuyNow={handleBuyNow} />
 
-      {/* Additional Info */}
-      <div>
-        <div className='text-xl font-bold text-primary space-y-1 flex items-center'>
-          <p className='mr-2'>Mô tả sản phẩm</p>
-          <button
-            onClick={() => setShowDescription((prev) => !prev)}
-            className='w-6 h-6 text-black flex items-center justify-center border border-gray-300 rounded-full bg-white hover:bg-gray-100 transition'
-            aria-label={showDescription ? 'Thu gọn' : 'Mở rộng'}
-          >
-            {showDescription ? '-' : '+'}
-          </button>
-        </div>
-        {showDescription && (
-          <div className='text-base text-primary mt-2'>
-            <h1>Thông tin sản phẩm và chính sách</h1>
-
-            <h2>Chất liệu</h2>
-            <ul className='list-disc pl-4'>
-              <li>Vải poly 2 da</li>
-            </ul>
-
-            <h2>Hướng dẫn bảo quản sản phẩm</h2>
-            <ul className='list-disc pl-4'>
-              <li>Giặt ở nhiệt độ bình thường, với đồ có màu tương tự.</li>
-              <li>Không dùng hóa chất tẩy lên sản phẩm.</li>
-              <li>Nên giặt tay và **LỘN NGƯỢC ÁO** trước khi giặt.</li>
-              <li>Không ủi trực tiếp lên hình in.</li>
-              <li>
-                Hạn chế sử dụng máy sấy và ủi (nếu có) chỉ nên ủi lên vải hoặc
-                sử dụng bàn ủi hơi nước ở nhiệt độ thích hợp.
-              </li>
-            </ul>
-
-            <h2>Chính sách đổi sản phẩm</h2>
-            <h3>1. Điều kiện đổi hàng</h3>
-            <ul className='list-disc pl-4'>
-              <li>Bạn lưu ý giữ lại hóa đơn để đổi hàng trong vòng 30 ngày.</li>
-              <li>
-                Đối với mặt hàng giảm giá, phụ kiện cá nhân (áo lót, khẩu trang,
-                vớ...) không nhận đổi hàng.
-              </li>
-              <li>
-                Tất cả sản phẩm đã mua sẽ không được đổi trả lại bằng tiền mặt.
-              </li>
-              <li>
-                Bạn có thể đổi size hoặc sản phẩm khác trong 30 ngày (Lưu ý: sản
-                phẩm chưa qua sử dụng, còn tag nhãn và hóa đơn mua hàng).
-              </li>
-              <li>
-                Bạn vui lòng gửi cho chúng mình clip đóng gói và hình ảnh của
-                đơn hàng đổi trả của bạn, nhân viên tư vấn sẽ xác nhận và tiến
-                hành lên đơn đổi trả cho bạn.
-              </li>
-            </ul>
-
-            <h3>2. Trường hợp khiếu nại</h3>
-            <ul className='list-disc pl-4'>
-              <li>Bạn phải có video unbox hàng.</li>
-              <li>Quay video rõ nét 6 mặt của gói hàng.</li>
-              <li>Quay rõ: Tên người nhận, mã đơn, địa chỉ, số điện thoại.</li>
-              <li>Clip không cắt ghép, chỉnh sửa.</li>
-            </ul>
-          </div>
-        )}
-      </div>
+      <ProductDescription />
     </div>
   );
 }
