+++ src/components/templates/(marketing)/Index/Product/Gallery.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useState, useEffect, useCallback, useMemo, useRef } from 'react';
 import Image from 'next/image';
 import {
   Carousel,
@@ -8,142 +7,34 @@ import {
   CarouselItem,
   CarouselNext,
   CarouselPrevious,
-  type CarouselApi,
 } from '@/components/ui/carousel';
-import { CategorySkeleton, Skeleton } from '@/components/ui/skeleton';
+import { CategorySkeleton } from '@/components/ui/skeleton';
 import { TProduct } from '@/types';
+import useGallery from '@/hooks/(product_details)/useGallery';
+import GalleryThumbnails from './(Gallery)/GalleryThumbnails';
 
 export default function Gallery({ product }: { product: TProduct }) {
-  //* Lưu instance của Carousel API để điều khiển carousel
-  const [api, setApi] = useState<CarouselApi>();
-
-  //* Lưu index của ảnh hiện tại đang được hiển thị trong carousel
-  const [current, setCurrent] = useState(0);
-
-  //* Mảng boolean, mỗi phần tử tương ứng với một ảnh, cho biết ảnh đó đã được load xong hay chưa
-  const [imagesLoaded, setImagesLoaded] = useState<boolean[]>([]);
-
-  //* Trạng thái auto play, true nếu carousel đang tự động chuyển slide
-  const [isAutoPlaying, setIsAutoPlaying] = useState(true);
-
-  //* Ref để đánh dấu đã preload ảnh hay chưa, tránh preload nhiều lần
-  const preloadedRef = useRef(false);
-
-  // Preload images only once
-  useEffect(() => {
-    if (preloadedRef.current) return;
-
-    const loadedStates = new Array(product.images?.length).fill(false);
-    setImagesLoaded(loadedStates);
-
-    product.images?.forEach((imgObj, index) => {
-      const img = new window.Image();
-      img.onload = () => {
-        setImagesLoaded((prev) => {
-          const newState = [...prev];
-          newState[index] = true;
-          return newState;
-        });
-      };
-      img.onerror = () => {
-        setImagesLoaded((prev) => {
-          const newState = [...prev];
-          newState[index] = true;
-          return newState;
-        });
-      };
-      img.src = imgObj?.imageUrl || '';
-    });
-
-    preloadedRef.current = true;
-  }, [product.images]);
-
-  // Set up carousel API
-  useEffect(() => {
-    if (!api) return;
-
-    setCurrent(api.selectedScrollSnap() + 1);
-
-    api.on('select', () => {
-      setCurrent(api.selectedScrollSnap() + 1);
-    });
-  }, [api]);
-
-  // Auto-play functionality
-  useEffect(() => {
-    if (!api || !isAutoPlaying) return;
-
-    const interval = setInterval(() => {
-      if (api.canScrollNext()) {
-        api.scrollNext();
-      } else {
-        api.scrollTo(0);
-      }
-    }, 5000);
-
-    return () => clearInterval(interval);
-  }, [api, isAutoPlaying]);
-
-  const handleImageSelect = useCallback(
-    (index: number) => {
-      if (api) {
-        api.scrollTo(index);
-      }
-    },
-    [api]
-  );
-
-  const handleMouseEnter = useCallback(() => {
-    setIsAutoPlaying(false);
-  }, []);
-
-  const handleMouseLeave = useCallback(() => {
-    setIsAutoPlaying(true);
-  }, []);
-
-  // Kiểm tra tất cả thumbnails đã load xong chưa
-  const allThumbnailsLoaded = useMemo(
-    () =>
-      product.images &&
-      imagesLoaded.length === product.images.length &&
-      imagesLoaded.every(Boolean),
-    [imagesLoaded, product.images]
-  );
+  const images = product.images || [];
+  const {
+    current,
+    imagesLoaded,
+    isAutoPlaying,
+    allThumbnailsLoaded,
+    setApi,
+    handleImageSelect,
+    handleMouseEnter,
+    handleMouseLeave,
+  } = useGallery(images);
 
   return (
     <div className='flex gap-4'>
       {/* Thumbnail Images */}
-      <div className='flex flex-col gap-2 min-w-[80px]'>
-        {!allThumbnailsLoaded
-          ? // Hiển thị skeleton placeholder cho toàn bộ thumbnails
-            Array.from({ length: product.images?.length || 4 }).map(
-              (_, idx) => (
-                <Skeleton key={idx} className='w-20 h-20 mb-2 rounded' />
-              )
-            )
-          : product.images?.map((image, index) => (
-              <button
-                key={index}
-                onClick={() => handleImageSelect(index)}
-                className={`w-20 h-20 bg-muted-foreground rounded overflow-hidden border-2 relative shadow-md hover:shadow-lg ${
-                  current === index + 1
-                    ? 'border-[#0e3bac] shadow-lg'
-                    : 'border-transparent'
-                } hover:border-[#0e3bac] transition-all duration-200`}
-              >
-                <Image
-                  src={image?.imageUrl || '/placeholder.svg'}
-                  alt={`Product view ${index + 1}`}
-                  width={80}
-                  height={80}
-                  className='w-full h-full object-cover'
-                  loading='lazy'
-                  sizes='80px'
-                  unoptimized
-                />
-              </button>
-            ))}
-      </div>
+      <GalleryThumbnails
+        images={images}
+        current={current}
+        allLoaded={allThumbnailsLoaded}
+        onSelect={handleImageSelect}
+      />
 
       {/* Main Image Carousel */}
       <div
@@ -153,7 +44,7 @@ export default function Gallery({ product }: { product: TProduct }) {
       >
         <Carousel setApi={setApi} className='w-full'>
           <CarouselContent>
-            {product.images?.map((image, index) => (
+            {images.map((image, index) => (
               <CarouselItem key={index}>
                 <div className='aspect-square bg-muted-foreground rounded-lg overflow-hidden relative'>
                   <Image
@@ -180,7 +71,7 @@ export default function Gallery({ product }: { product: TProduct }) {
 
           {/* Image Indicators */}
           <div className='absolute bottom-4 left-1/2 -translate-x-1/2 flex gap-2'>
-            {product.images?.map((_, index) => (
+            {images.map((_, index) => (
               <button
                 key={index}
                 onClick={() => handleImageSelect(index)}
